+++ src/main/java/nextstep/ladder/LadderMain.java
@@ -1,4 +1,20 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.LadderGame;
+import nextstep.ladder.domain.Person;
+import nextstep.ladder.view.InputView;
+import nextstep.ladder.view.OutputView;
+
+import java.util.List;
+
 public class LadderMain {
+
+    public static void main(String[] args) {
+        List<Person> names = InputView.getPersonNames();
+        int countOfPerson = names.size();
+        int height = InputView.getHeight();
+
+        OutputView.printLadder(height, names, new LadderGame(height, countOfPerson));
+    }
+
 }
